notion list

+++ lib/review/ReviewCounter.php
@@ -264,9 +264,7 @@ class ReviewCounter
 				$rating += $item[ 'value' ];
 			}
 
-			// $rating = number_format( ( float ) ( $rating / $amount ), 1, '.', '');
-			
-			$rating = number_format( ( int ) ( $rating / $amount ), 1, '.', '');
+			$rating = number_format( ( float ) ( $rating / $amount ), 1, '.', '');
 		}
 
 		if ( $amount == 5 )
